JG-2007-11-26-0: Bugfix:    Fixed minor memory leaks.                  Review:    EB Issue number: Obtained from: Submitted by: Reviewed by:

+++ source/clients/common/sge_cqueue_qconf.c
@@ -332,8 +332,10 @@ cqueue_provide_modify_context(sge_gdi_ctx_class_t *ctx, lListElem **this_elem, l
                                                      false);
       
       if (answer_list_output(answer_list)) {
-         unlink(filename);
-         FREE(filename);
+         if (filename != NULL) { 
+            unlink(filename);
+            FREE(filename);
+         }
          DRETURN(false);
       }
  
@@ -792,7 +794,7 @@ cqueue_list_sick(sge_gdi_ctx_class_t *ctx, lList **answer_list)
 
       if (sge_dstring_get_string(&ds)) {
          printf(sge_dstring_get_string(&ds));
- 	      ret = false;
+         ret = false;
       }
 
       sge_dstring_free(&ds);
@@ -910,10 +912,10 @@ static int write_QU_consumable_config_list(const lListElem *ep, int nm,
          }
       }
    }
-   
+
    /* for CR 6433628, adding NONE string when there are no complex values */
    if (!has_elems){
       sge_dstring_append(buffer, NONE_STR);
-   }  
+   }
    DRETURN(1);
 }
